fix(react): duplicate events being fired in ionic/react (#17321)
