+++ react/src/components/utils.ts
@@ -1,17 +1,32 @@
+/**
+ * Checks if an event is supported in the current execution environment.
+ * @license Modernizr 3.0.0pre (Custom Build) | MIT
+ */
+function isCoveredByReact(eventNameSuffix: string) {
+  const eventName = 'on' + eventNameSuffix;
+  let isSupported = eventName in document;
+
+  if (!isSupported) {
+    const element = document.createElement('div');
+    element.setAttribute(eventName, 'return;');
+    isSupported = typeof (<any>element)[eventName] === 'function';
+  }
+
+  return isSupported;
+}
 
 function syncEvent(node: Element, eventName: string, newEventHandler: (e: Event) => any) {
-  const eventNameLc = eventName[0].toLowerCase() + eventName.substring(1);
   const eventStore = (node as any).__events || ((node as any).__events = {});
-  const oldEventHandler = eventStore[eventNameLc];
+  const oldEventHandler = eventStore[eventName];
 
   // Remove old listener so they don't double up.
   if (oldEventHandler) {
-    node.removeEventListener(eventNameLc, oldEventHandler);
+    node.removeEventListener(eventName, oldEventHandler);
   }
 
   // Bind new listener.
   if (newEventHandler) {
-    node.addEventListener(eventNameLc, eventStore[eventNameLc] = function handler(e: Event) {
+    node.addEventListener(eventName, eventStore[eventName] = function handler(e: Event) {
       newEventHandler.call(this, e);
     });
   }
@@ -35,7 +50,12 @@ export function attachEventProps<E extends HTMLElement>(node: E, props: any) {
     }
 
     if (name.indexOf('on') === 0 && name[2] === name[2].toUpperCase()) {
-      syncEvent(node, name.substring(2), props[name]);
+      const eventName = name.substring(2);
+      const eventNameLc = eventName[0].toLowerCase() + eventName.substring(1);
+
+      if (!isCoveredByReact(eventNameLc)) {
+        syncEvent(node, eventNameLc, props[name]);
+      }
     } else {
       (node as any)[name] = props[name];
     }
